Enum bincode_core::DeserializeError

pub enum DeserializeError<'a, R: CoreRead<'a>> {
    Read(R::Error),
    InvalidBoolValue(u8),
    InvalidCharEncoding,
    Utf8(Utf8Error),
    InvalidOptionValue(u8),
    LimitError(LimitError),
    InvalidCast {
        from_type: &'static str,
        to_type: &'static str,
    },
    InvalidUtf8Encoding(Utf8Error),
    InvalidValueRange,
    ExtensionPoint,
}

Errors that can occur while deserializing

Variants

Read(R::Error)

Failed to read from the provided CoreRead. The inner exception is given.

InvalidBoolValue(u8)

Invalid bool value. Only 0 and 1 are valid values.

InvalidCharEncoding

Invalid character encoding while trying to deserialize a &str.

Utf8(Utf8Error)

UTF8 error while trying to deserialize a &str

InvalidOptionValue(u8)

Invalid value for the Option part of Option<T>. Only 0 and 1 are accepted values.

LimitError(LimitError)

Limit error reached. See the inner exception for more info.

InvalidCast

Could not cast from type from_type to type to_type. Usually this means that the data is encoded with a different version or protocol.

Fields of InvalidCast

from_type: &'static str

The base type that was being casted from

to_type: &'static str

The target type that was being casted to

InvalidUtf8Encoding(Utf8Error)

Invalid UTF8 encoding while trying to parse a &str or String

InvalidValueRange

Invalid value (u128 range): you may have a version or configuration disagreement?

ExtensionPoint

Byte 255 is treated as an extension point; it should not be encoding anything. Do you have a mismatched bincode version or configuration?